refactored most types into timelineTypes to avoid having types everywhere
