--- src/components/video-editor/timeline/hooks/timelineHookTypes.ts
@@ -1,9 +1,0 @@
-export interface RegionIdLite {
-	id: string;
-	startMs: number;
-	endMs: number;
-}
-
-export interface AudioRegionLite extends RegionIdLite {
-	trackIndex?: number;
-}
